+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New `assistant_message()` solver.
 - Prompt templates: Ignore template placeholders that don't map to passed parameters in `prompt_template()`, and system/user/assistnat solvers.
 - Google: Handle system messages with content lists and input with system but no user messages.
+- Google: Ensure that a completion choice is provided even when none are returned by the service.
 - Inspect View: Improve the display of subtasks with no inputs or events.
 - Inspect View: Fix transcript display of phantom subtask or other phantom events.
 - Inspect View: Fix formatting issues in sample error display