+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v0.3.67 (18 February 2025)
 
 - Memoize calls to `get_model()` so that model instances with the same parameters are cached and re-used (pass `memoize=False` to disable).
 - Async context manager for `Model` class for optional scoped usage of model clients.
@@ -12,7 +12,6 @@
 - Bugfix: Raise error for empty dataset (rather than providing a dummy sample).
 - Bugfix: Specify markup=False for textual static controls (stricter parser in textual 2.0 leading to exceptions).
 - Bugfix: Temporarily pin to textual==1.0.0 while they chase all of their regressions in 2.0
-
 
 ## v0.3.66 (17 February 2025)